Added add_products.py : - The script emulates the process of saving scraped data, such as EAN and seller name, into the database.
irrelevant improvements at admin.py

index.html is showing the data from the Product table from the database.

views.py is retrieving all the data from the Product table in the database, checks if the data is correctly retrieved then execute the SQL query.

The migrations folder is filled with test data for the PRODUCT table, primarily for testing purposes. These files contain the data that will be inserted into the database through schema.sql.